refactor: changes in project and model structure

+++ src/main/java/com/local/model/Payment.java
@@ -1,16 +1,17 @@
 package com.local.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class Payment implements Comparable<Payment>{
     private int id;
     private User user;
     private Cart cart;
-    private double amount;
+    private BigDecimal amount;
     private LocalDateTime lastUpdate;
     private PaymentStatus status;
 
-    public Payment(int id, User user, Cart cart, double amount, LocalDateTime lastUpdate, PaymentStatus status) {
+    public Payment(int id, User user, Cart cart, BigDecimal amount, LocalDateTime lastUpdate, PaymentStatus status) {
         this.id = id;
         this.user = user;
         this.cart = cart;
@@ -51,11 +52,11 @@ public class Payment implements Comparable<Payment>{
         this.cart = cart;
     }
 
-    public double getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(double amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
